backend/requirements.txt
fastapi
uvicorn
pydantic
torch
transformers
pdfplumber
motor
passlib
python-jose[cryptography]
